maiya, engine: no library division in a frame's work, lighter HUD text; 52-55 fps
A 32-bit or signed division on the 68000 goes through the library, several
times the cost of one divu.w. Maiya's frame had them in the front plane's
wrap (three a frame), the enemy animation phase (one per creature), the
spark shrink, the health bars, the walk cycle and the guardians' timers;
each is now a 16-bit divide or a subtraction with the same result. HUD
numbers take a digit at a time from two word divides (mg_div10).

Shots and pickups are flushed instead of re-uploaded every frame (their
tiles only change through the setters), and the road ahead is scanned for
spawns every other frame.

Engine: ng_fix_putc writes its cell's map word itself instead of going
through fixtext_out, which measured a one-character string first (about
half the cost of a character). ng_palette_fx's flash and pulse divides
are 16-bit (the products fit a word). helloworld keeps its size, demo is
4 bytes smaller; demo_plus is unchanged.

Measured in MAME, walk and fight through stages 1, 2, 4 and 6: play
frames average 12.5-14.1 ms, 52-55 fps; 7-14% of play frames still
overrun (a hint line or a burst of new sprites). Guardian fights 53-54 fps;
the warp's palette effect about 40.

+++ sdk/2d_engine/ng_palette_fx.c
@@ -523,7 +523,7 @@ void NEOGEO_USER ng_palette_fx_update(void)
             /* Blend from base towards white; intensity decays over duration */
             uint8_t blend;
 
-            blend = (uint8_t)(((uint16_t)s->timer * 255u) / (uint16_t)s->duration);
+            blend = (uint8_t)((uint16_t)((uint16_t)s->timer * 255u) / (uint16_t)s->duration);
             /* White: R=31, G=31, B=31 → pass as scaled 0..255: 255 */
             palfx_blend_to(s->work_pal, s->base_pal, 255, 255, 255, blend);
             palfx_put(s->palette_slot, s->work_pal);
@@ -540,7 +540,7 @@ void NEOGEO_USER ng_palette_fx_update(void)
         case NG_PALFX_FLASH_RED: {
             uint8_t blend;
 
-            blend = (uint8_t)(((uint16_t)s->timer * 200u) / (uint16_t)s->duration);
+            blend = (uint8_t)((uint16_t)((uint16_t)s->timer * 200u) / (uint16_t)s->duration);
             palfx_blend_to(s->work_pal, s->base_pal, 255, 0, 0, blend);
             palfx_put(s->palette_slot, s->work_pal);
 
@@ -556,7 +556,7 @@ void NEOGEO_USER ng_palette_fx_update(void)
         case NG_PALFX_FLASH_BLUE: {
             uint8_t blend;
 
-            blend = (uint8_t)(((uint16_t)s->timer * 180u) / (uint16_t)s->duration);
+            blend = (uint8_t)((uint16_t)((uint16_t)s->timer * 180u) / (uint16_t)s->duration);
             palfx_blend_to(s->work_pal, s->base_pal, 0, 64, 255, blend);
             palfx_put(s->palette_slot, s->work_pal);
 
@@ -584,9 +584,9 @@ void NEOGEO_USER ng_palette_fx_update(void)
 
             /* Triangle wave: ramp up then ramp down */
             if (s->timer < half) {
-                phase = (uint8_t)((uint16_t)s->timer * 127u / (uint16_t)half);
+                phase = (uint8_t)((uint16_t)((uint16_t)s->timer * 127u) / (uint16_t)half);
             } else {
-                phase = (uint8_t)(127u - (uint16_t)(s->timer - half) * 127u / (uint16_t)(s->duration - half));
+                phase = (uint8_t)(127u - (uint16_t)((uint16_t)(s->timer - half) * 127u) / (uint16_t)(s->duration - half));
             }
             brightness = (uint8_t)(128 + phase);
